+++ backend/api/src/main/java/com/yat2/episode/auth/oauth/KakaoIdTokenVerifier.java
@@ -7,7 +7,6 @@ import com.nimbusds.jose.util.DefaultResourceRetriever;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.proc.ConfigurableJWTProcessor;
 import com.nimbusds.jwt.proc.DefaultJWTProcessor;
-import com.yat2.episode.auth.KakaoProperties;
 import org.springframework.stereotype.Component;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.jwk.source.RemoteJWKSet;
